app/requirements.txt
debugpy
Django
django-cors-headers
django-extensions
django-filter
django-lifecycle
djangorestframework
djangorestframework_simplejwt
drf-spectacular[sidecar]
gunicorn
psycopg
psycopg-binary
psycopg-pool
psycopg2-binary
rdflib
requests
tzdata
websocket-client
yq
